+++ src/constants/messages.ts
@@ -5,6 +5,7 @@ export const API = {
   getAlertRulesFailed: 'Get Alert Rules: API Request failed. See Console for details.',
   getAnnotationsFailed: 'Get Annotations: API Request failed. See Console for details.',
   getDataSourcesFailed: 'Get Data Sources: API Request failed. See Console for details.',
+  getDashboardsMetaFailed: 'Get Dashboards Meta: API Request failed. See Console for details.',
   getHealthFailed: 'Get Health: API Request failed. See Console for details.',
   getOrgFailed: 'Get Org: API Request failed. See Console for details.',
   getUsersFailed: 'Get Users: API Request failed. See Console for details.',
@@ -24,5 +25,6 @@ export const MESSAGES = {
   requestHealth: 'Returns health information',
   requestNone: 'For internal API calls',
   requestOrgUsers: 'Returns all users within the current organization',
+  dashboardsMeta: 'Returns dashboards metadata',
   version: 'Version',
 };